Code clean-up

+++ Roy-T.AStar/Grid.cs
@@ -5,13 +5,13 @@ namespace RoyT.AStar
 {
     /// <summary>
     /// Representation of your world for the pathfinding algorithm.
-    /// Use SetCellCost change the cost of traversing a cell
+    /// Use SetCellCost to change the cost of traversing a cell.
     /// Use BlockCell to make a cell completely intraversable.
     /// </summary>
     public sealed class Grid
     {       
-        private readonly double DefaultCost;
-        private readonly double[] Weights;        
+        private readonly float DefaultCost;
+        private readonly float[] Weights;        
 
         /// <summary>
         /// Creates a grid
@@ -19,7 +19,7 @@ namespace RoyT.AStar
         /// <param name="dimX">The x-dimension of your world</param>
         /// <param name="dimY">The y-dimesion of your world</param>
         /// <param name="defaultCost">The default cost every cell is initialized with</param>
-        public Grid(int dimX, int dimY, double defaultCost)
+        public Grid(int dimX, int dimY, float defaultCost)
         {
             if (defaultCost <= 0)
             {
@@ -28,7 +28,7 @@ namespace RoyT.AStar
             }
 
             this.DefaultCost = defaultCost;
-            this.Weights = new double[dimX * dimY];
+            this.Weights = new float[dimX * dimY];
             this.DimX = dimX;
             this.DimY = dimY;
 
@@ -53,7 +53,7 @@ namespace RoyT.AStar
         /// </summary>
         /// <param name="position">A position inside the grid</param>
         /// <param name="cost">The cost of traversing the cell, always larger than zero</param>
-        public void SetCellCost(Position position, double cost)
+        public void SetCellCost(Position position, float cost)
         {
             if (cost < 0)
             {
@@ -65,19 +65,20 @@ namespace RoyT.AStar
         }
 
         /// <summary>
-        /// Makes the cell inaccessible
+        /// Makes the cell intraversable
         /// </summary>
         /// <param name="position">A position inside the grid</param>
-        public void BlockCell(Position position) => SetCellCost(position, double.PositiveInfinity);
+        public void BlockCell(Position position) => SetCellCost(position, float.PositiveInfinity);
 
         /// <summary>
-        /// Makes the cell accessible, gives it the default traversal cost as provided in the constructor
+        /// Makes the cell traversable, gives it the default traversal cost as provided in the constructor
         /// </summary>
         /// <param name="position">A position inside the grid</param>
         public void UnblockCell(Position position) => SetCellCost(position, this.DefaultCost);
 
         /// <summary>
-        /// Looks-up the cost for traversing a given cell
+        /// Looks-up the cost for traversing a given cell, if a cell is blocked (<see cref="BlockCell"/>) 
+        /// +infinity is returned
         /// </summary>
         /// <param name="position">A position inside the grid</param>
         /// <returns>The cost</returns>
@@ -92,13 +93,13 @@ namespace RoyT.AStar
         /// </summary>
         /// <param name="position">A position inside the grid</param>
         /// <returns>The cost</returns>
-        internal double GetCellCostUnchecked(Position position)
+        internal float GetCellCostUnchecked(Position position)
         {
             return this.Weights[this.DimX * position.X + position.Y];
         }
 
         /// <summary>
-        /// Computs lowest-cost path from start to end inside the grid for an agent that can
+        /// Computes the lowest-cost path from start to end inside the grid for an agent that can
         /// move both diagonal and lateral
         /// </summary>
         /// <param name="start">The start position</param>
@@ -108,12 +109,12 @@ namespace RoyT.AStar
             => GetPath(start, end, RangeOfMotion.Full);
 
         /// <summary>
-        /// Computs lowest-cost path from start to end inside the grid for an agent with a custom
+        /// Computes the lowest-cost path from start to end inside the grid for an agent with a custom
         /// range of motion
         /// </summary>
         /// <param name="start">The start position</param>
         /// <param name="end">The end position</param>
-        /// <param name="rangeOfMotion">The range of motion of the agent, <see cref="RangeOfMotion"/> for several built in options</param>
+        /// <param name="rangeOfMotion">The range of motion of the agent, <see cref="RangeOfMotion"/> for several built-in options</param>
         /// <returns>positions of cells, from start to end, on the shortest path from start to end</returns>
         public IList<Position> GetPath(Position start, Position end, Offset[] rangeOfMotion)
         {
